requirements.txt
alembic==1.13.1
attrs==21.4.0
blinker==1.8.2
click==8.1.7
Faker==12.0.1
flake8==4.0.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.1.6
MarkupSafe==2.1.5
mccabe==0.6.1
mixer==7.2.2
mypy==1.10.0
mypy-extensions==1.0.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.1
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.0
Werkzeug==3.0.3
WTForms==3.0.1
